switched fetching the mock data to be from the backend. There is a slight typing incompatibility, but we'll ignore it since we'll seed the database net
