+++ server/routers/_app.ts
@@ -1,5 +1,6 @@
 import { router, publicProcedure } from '../trpc';
 import { prisma } from '../prisma';
+import { personRouter } from './person';
 
 export const appRouter = router({
   healthcheck: publicProcedure.query(async (req) => {
@@ -7,6 +8,7 @@ export const appRouter = router({
 
     return { greeting: 'hello world' };
   }),
+  person: personRouter,
 });
 
 export type AppRouter = typeof appRouter;
